pyproject.toml
[tool.poetry]
name = "lmd-ukr"
version = "0.1.0"
description = "Out of curiosity, comments activity on Le Monde Ukraine coverage"
authors = ["matthieuvion <vionmatthieu@gmail.com>"]
readme = "README.md"
packages = [{include = "lmd_ukr"}]

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.28.2"
jupyterlab = "^3.6.1"
httpx = "^0.23.3"
python-dotenv = "^0.21.1"
selectolax = "^0.3.12"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
